refactor(tests): split refresh contract modules
